# rudi-microservice/rudi-microservice-acl/openapi/rudi-acl-api.json
{
	"openapi": "3.0.0",
	"info": {
		"description": "API Public ACL Rudi",
		"version": "0.0.1",
		"title": "rudi-acl-api",
		"termsOfService": "http://swagger.io/terms/",
		"contact": {
			"email": "contact@rudi.fr"
		},
		"license": {
			"name": "Apache 2.0",
			"url": "http://www.apache.org/licenses/LICENSE-2.0.html"
		}
	},
	"servers": [
		{
			"url": "/{basePath}",
			"description": "Rudi Acl API",
			"variables": {
				"basePath": {
					"default": "acl/v1"
				}
			}
		}
	],
...
	"paths": {
...
		"/users": {
			"get": {
				"tags": [
					"acl"
				],
				"summary": "Collection d'utilisateurs",
				"description": "Retourne une liste d'utilisateurs'",
				"operationId": "searchUsers",
				"parameters": [
					{
						"in": "query",
						"name": "login",
						"description": "Tout ou partie du login de l'utilisateur",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "password",
						"description": "Mot-de-passe (en clair) de l'utilisateur",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "lastname",
						"description": "Tout ou partie du nom de famille de l'utilisateur",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "firstname",
						"description": "Tout ou partie du prénom de famille de l'utilisateur",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "company",
						"description": "Tout ou partie du prénom de la société de l'utilisateur",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "type",
						"description": "Type d 'utilisateur",
						"required": false,
						"schema": {
							"$ref": "rudi-acl-model.json#/components/schemas/UserType"
						}
					},
					{
						"in": "query",
						"name": "role-uuids",
						"description": "UUIDS des roles",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "string",
								"format": "uuid"
							}
						}
					},
					{
						"in": "query",
						"name": "user-uuids",
						"description": "UUIDS des users",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "string",
								"format": "uuid"
							}
						}
					},
					{
						"in": "query",
						"name": "login-and-denomination",
						"description": "Tout ou partie du login, du nom ou du prenom de l'utilisateur",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "offset",
						"description": "Index de début",
						"required": false,
						"schema": {
							"type": "integer",
							"format": "int32"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"description": "Nombre de résultats à retourner",
						"required": false,
						"schema": {
							"type": "integer",
							"format": "int32"
						}
					},
					{
						"in": "query",
						"name": "order",
						"required": false,
						"schema": {
							"type": "string",
							"description": "liste des tris à réaliser. Exemple :  login,lastname"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "rudi-acl-model.json#/components/schemas/UserPageResult"
								}
							}
						}
					},
					"500": {
						"description": "Internal server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						}
					}
				}
			},
			"post": {
				"tags": [
					"acl"
				],
				"summary": "Ajout d'un nouvel utilisateur",
				"description": "Permet d'ajouter un nouvel utilisateur",
				"operationId": "createUser",
				"requestBody": {
					"description": "info utilisateur",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "rudi-acl-model.json#/components/schemas/User"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "rudi-acl-model.json#/components/schemas/User"
								}
							}
						}
					},
					"500": {
						"description": "Internal server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						}
					}
				}
			},
			"put": {
				"tags": [
					"acl"
				],
				"summary": "Mise à jour d'un utilisateur",
				"description": "Mise à jour d'un utilisateur",
				"operationId": "updateUser",
				"requestBody": {
					"description": "info utilisateur",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "rudi-acl-model.json#/components/schemas/User"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "rudi-acl-model.json#/components/schemas/User"
								}
							}
						}
					},
					"500": {
						"description": "Internal server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						}
					}
				}
			}
		},
		"/users/{userUuid}": {
			"get": {
				"tags": [
					"acl"
				],
				"summary": "Retourne un utilisateur",
				"description": "Retourne un utilisateur",
				"operationId": "getUser",
				"parameters": [
					{
						"name": "userUuid",
						"in": "path",
						"description": "uuid pour la récupération d'un élément",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "rudi-acl-model.json#/components/schemas/User"
								}
							}
						}
					},
					"500": {
						"description": "Internal server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						}
					}
				}
			},
			"delete": {
				"tags": [
					"acl"
				],
				"summary": "Suppression d'un utilisateur",
				"description": "Suppression d'un utilisateur",
				"operationId": "deleteUser",
				"parameters": [
					{
						"name": "userUuid",
						"in": "path",
						"description": "info uuid",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK"
					},
					"500": {
						"description": "Internal server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						}
					}
				}
			}
		},
		"/users/count": {
			"get": {
				"tags": [
					"acl"
				],
				"summary": "Nombre d'utilisateurs",
				"description": "Retourne le nombre d'utilisateurs de type USER",
				"operationId": "countUsers",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"type": "integer",
									"format": "int64"
								}
							}
						}
					},
					"500": {
						"description": "Internal server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						}
					}
				}
			}
		},
...
		"/users/{userUuid}/addresses": {
			"get": {
				"tags": [
					"acl"
				],
				"summary": "Collection d'adresses",
				"description": "Retourne les adresses d'un utilisateur'",
				"operationId": "getAddresses",
				"parameters": [
					{
						"name": "userUuid",
						"in": "path",
						"description": "Uuid de l''utilisateur",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "rudi-acl-model.json#/components/schemas/AbstractAddress"
									}
								}
							}
						}
					},
					"500": {
						"description": "Internal server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						}
					}
				}
			},
			"post": {
				"tags": [
					"acl"
				],
				"summary": "Ajout d'une adresse",
				"description": "Ajout d'une adresse",
				"operationId": "createAddress",
				"parameters": [
					{
						"name": "userUuid",
						"in": "path",
						"description": "Uuid de l'utilisateur'",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					}
				],
				"requestBody": {
					"description": "info address",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "rudi-acl-model.json#/components/schemas/AbstractAddress"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "rudi-acl-model.json#/components/schemas/AbstractAddress"
								}
							}
						}
					},
					"500": {
						"description": "Internal server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						}
					}
				}
			},
			"put": {
				"tags": [
					"acl"
				],
				"summary": "Mise à jour d'une adresse",
				"description": "Mise à jour d'une adresse d'un utilisateur'",
				"operationId": "updateAddress",
				"parameters": [
					{
						"name": "userUuid",
						"in": "path",
						"description": "Uuid de l'utilisateur'",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					}
				],
				"requestBody": {
					"description": "info address",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "rudi-acl-model.json#/components/schemas/AbstractAddress"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "rudi-acl-model.json#/components/schemas/AbstractAddress"
								}
							}
						}
					},
					"500": {
						"description": "Internal server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						}
					}
				}
			}
		},
		"/users/{userUuid}/addresses/{addressUuid}": {
			"get": {
				"tags": [
					"acl"
				],
				"summary": "Retourne une adresse",
				"description": "Retourne une adresse d'un utilisateur'",
				"operationId": "getAddress",
				"parameters": [
					{
						"name": "userUuid",
						"in": "path",
						"description": "uuid de l'utilisateur",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					},
					{
						"name": "addressUuid",
						"in": "path",
						"description": "uuid de l'adresse'",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "rudi-acl-model.json#/components/schemas/AbstractAddress"
								}
							}
						}
					},
					"500": {
						"description": "Internal server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						}
					}
				}
			},
			"delete": {
				"tags": [
					"acl"
				],
				"summary": "Suppression d'une adresse",
				"description": "Suppression d'une adresse d'un utilisateur'",
				"operationId": "deleteAddress",
				"parameters": [
					{
						"name": "userUuid",
						"in": "path",
						"description": "uuid de l'utilisateur'",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					},
					{
						"name": "addressUuid",
						"in": "path",
						"description": "uuid de l''adresse",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK"
					},
					"500": {
						"description": "Internal server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						}
					}
				}
			}
		},
...
		"/users/{login}/password": {
			"put": {
				"tags": [
					"acl"
				],
				"summary": "Gère la modification du mot de passe d'un utilisateur d'organisation",
				"description": "permet la modification du mot de passe d'un utilisateur d'organisation",
				"operationId": "updateUserPassword",
				"parameters": [
					{
						"name": "login",
						"in": "path",
						"description": "Login de l'utilisateur d'organisation'",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "Password du user",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "./rudi-acl-model.json#/components/schemas/PasswordUpdate"
							}
						}
					}
				},
				"responses": {
					"204": {
						"description": "OK"
					},
					"500": {
						"description": "Internal server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						}
					}
				}
			}
		},
...
	},
	"components": {
		"schemas": {
			"ApiError": {
				"type": "object",
				"properties": {
					"code": {
						"type": "string"
					},
					"label": {
						"type": "string"
					}
				}
			},
...
		}
	}
}